+++ src/main/java/dev/repository/ChauffeurRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import dev.domain.Chauffeur;
 
-public interface ChauffeurRepo extends JpaRepository<Chauffeur, Integer> {
+public interface ChauffeurRepo extends JpaRepository<Chauffeur, Long> {
 
 }
